(react) change: move folders to src
