+++ expenses-react/src/App.js
@@ -1,17 +1,17 @@
 import logo from './logo.svg';
-import List from "./js/components/List";
-import Form from "./js/components/Form";
+import List from "./components/List";
+import Form from "./components/Form";
 
 function App() {
   return (
     <>
-      <div className="App">
-        <h2>Expenses</h2>
-        <List />
-      </div>
       <div>
         <h2>Add a new article</h2>
         <Form />
+      </div>
+      <div className="App">
+        <h2>Expenses</h2>
+        <List />
       </div>
     </>
   );
