+++ expenses-react/src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from "react-dom";
 import { Provider } from "react-redux";
-import store from "./js/store/index"
+import store from "./store/index"
 import App from './App';
 
 render(
